requirements.txt
click==8.0.4
pdf2image==1.16.0
Pillow==9.0.1
poppler-utils==0.1.0
tk==0.1.0
